Get product and subcategories, include breadcrumbs
